Use bit-ssb project template
