+++ Bit.BlazorUIPlayground.sln
@@ -1,0 +1,31 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31611.283
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{5CF43F76-BB71-4B5B-B4DF-1C753E042A8F}"
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		.\src\global.json = .\src\global.json
+		.\src\Clean.bat = .\src\Clean.bat
+	EndProjectSection
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = ".\src\Bit.BlazorUIPlayground.Web", ".\src\Bit.BlazorUIPlayground.Web\Bit.BlazorUIPlayground.Web.csproj", "{2B232F0F-9704-451E-8E3E-313AE0BEEF03}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{2B232F0F-9704-451E-8E3E-313AE0BEEF03}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2B232F0F-9704-451E-8E3E-313AE0BEEF03}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2B232F0F-9704-451E-8E3E-313AE0BEEF03}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2B232F0F-9704-451E-8E3E-313AE0BEEF03}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {61F7FB11-1E47-470C-91E2-47F8143E1572}
+	EndGlobalSection
+EndGlobal
