+++ README.md
@@ -4,6 +4,6 @@ With Gitpod, you can access the ultimate online development platform, right in y
 
 Experience lightning-fast speeds and seamless integration as you dive into the world of Bit BlazorUI, and elevate your web development game to new heights.
 
-Simply naviate to the `src/BitBlazorUIPlayground.Web` folder and run the project with `dotnet run` command!
+Simply naviate to the `src/Bit.BlazorUIPlayground.Web` folder and run the project with `dotnet run` command!
 
 [![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/#https://github.com/bitfoundation/bit-blazorui-playground)
