+++ src/Bit.BlazorUIPlayground.Web/Bit.BlazorUIPlayground.Web.csproj
@@ -1,0 +1,26 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <Nullable>enable</Nullable>
+        <ImplicitUsings>enable</ImplicitUsings>
+    </PropertyGroup>
+
+    <ItemGroup>
+        <PackageReference Include="Bit.BlazorUI" Version="4.9.4-pre-01" />
+        
+        <PackageReference Include="Bit.CodeAnalyzers" Version="4.9.4-pre-01">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+        
+        <PackageReference Include="Bit.SourceGenerators" Version="4.9.4-pre-01">
+            <PrivateAssets>all</PrivateAssets>
+            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+        </PackageReference>
+
+        <Using Include="Microsoft.JSInterop" />
+        <Using Include="Bit.BlazorUIPlayground.Web.Shared" />
+    </ItemGroup>
+
+</Project>
