+++ src/Bit.BlazorUIPlayground.Web/Pages/About.razor
@@ -1,0 +1,62 @@
+﻿@page "/about"
+
+<PageTitle>About</PageTitle>
+
+<div class="page">
+    <section class="bitplatform">
+        <BitLink Href="https://bitplatform.dev" Target="_blank">
+            <h2>bitplatform</h2>
+        </BitLink>
+    </section>
+
+    <section class="page-section vision-section">
+        <div class="page-section-contnet">
+            <h1 class="vision-title">
+                Our vision is to empower organizations and people to maximize their performance through <span class="text--gradiant">high quality development services and products</span>.
+            </h1>
+        </div>
+    </section>
+
+    <section class="page-section values-section">
+        <div class="page-section-contnet">
+            <h2 class="section-title">Our values</h2>
+            <div class="card-list">
+                <div class="value-card">
+                    <div class="value-card-title quality-icon">
+                        Quality
+                    </div>
+                    <div class="value-card-desc">
+                        We expertly execute our work, utilizing innovation and ingenuity to provide high-quality services and products.
+                    </div>
+                </div>
+
+                <div class="value-card">
+                    <div class="value-card-title integrity-icon">
+                        Integrity
+                    </div>
+                    <div class="value-card-desc">
+                        Conduct our business in a legal and ethical manner, build lasting relationships through honesty, respect and trust with our employees as well as our business partners.
+                    </div>
+                </div>
+
+                <div class="value-card">
+                    <div class="value-card-title transparency-icon">
+                        Transparency
+                    </div>
+                    <div class="value-card-desc">
+                        Being honest and open when communicating with our employees and business partners.
+                    </div>
+                </div>
+            </div>
+        </div>
+    </section>
+
+    <section class="page-section download-section">
+        <div class="page-section-contnet">
+            <div class="download-container">
+                <div class="download-value">2.7M</div>
+                <div class="download-name">Total downloads of packages</div>
+            </div>
+        </div>
+    </section>
+</div>
