+++ src/Bit.BlazorUIPlayground.Web/Pages/_Host.cshtml
@@ -1,0 +1,29 @@
+﻿@page "/"
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
+@using Microsoft.AspNetCore.Components.Web
+@namespace Bit.BlazorUIPlayground.Web.Pages
+
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <base href="~/" />
+    <meta charset="utf-8" />
+    <link rel="icon" type="image/x-icon" href="favicon.ico" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="_content/Bit.BlazorUI/styles/bit.blazorui.min.css" />
+    <link rel="stylesheet" href="Bit.BlazorUIPlayground.Web.styles.css" />
+    <component type="typeof(HeadOutlet)" render-mode="Server" />
+    <style>
+        html, body {
+            margin: 0;
+            padding: 0;
+            height: 100vh;
+        }
+    </style>
+</head>
+<body>
+    <component type="typeof(App)" render-mode="Server" />
+    <script src="_framework/blazor.server.js"></script>
+    <script src="_content/Bit.BlazorUI/scripts/bit.blazorui.min.js"></script>
+</body>
+</html>
