+++ src/Bit.BlazorUIPlayground.Web/Properties/launchSettings.json
@@ -1,0 +1,13 @@
+{
+    "profiles": {
+        "Bit.BlazorUIPlayground.Web": {
+            "commandName": "Project",
+            "dotnetRunMessages": true,
+            "launchBrowser": true,
+            "environmentVariables": {
+                "ASPNETCORE_ENVIRONMENT": "Development"
+            },
+            "applicationUrl": "https://localhost:4001;http://localhost:4000"
+        }
+    }
+}
