+++ src/Bit.BlazorUIPlayground.Web/Shared/Footer.razor
@@ -1,0 +1,5 @@
+﻿<footer>
+    <section>
+        <BitLabel Class="copyright">Copyright © 2023 Bit.BlazorUIPlayground | All Rights Reserved</BitLabel>
+    </section>
+</footer>
