+++ src/Bit.BlazorUIPlayground.Web/Shared/MainLayout.razor
@@ -1,0 +1,11 @@
+﻿@inherits LayoutComponentBase
+
+<Header />
+
+<main>
+    <article>
+        @Body
+    </article>
+
+    <Footer />
+</main>
