+++ src/Bit.BlazorUIPlayground.Web/Shared/MainLayout.razor.css
@@ -1,0 +1,14 @@
+﻿::deep *, * {
+    box-sizing: border-box;
+    font-family: 'Lucida Sans', 'Lucida Sans Regular', 'Lucida Grande', 'Lucida Sans Unicode', Geneva, Verdana, sans-serif;
+}
+
+main {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+}
+
+article {
+    flex-grow: 1;
+}
