+++ src/Bit.BlazorUIPlayground.Web/_Imports.razor
@@ -1,0 +1,8 @@
+﻿@using System.Net.Http
+@using Microsoft.JSInterop
+@using Microsoft.AspNetCore.Components.Web
+@using Microsoft.AspNetCore.Components.Routing
+@using Microsoft.AspNetCore.Components.Web.Virtualization
+@using Bit.BlazorUI
+@using Bit.BlazorUIPlayground.Web
+@using Bit.BlazorUIPlayground.Web.Shared
