+++ src/Bit.BlazorUIPlayground.Web/appsettings.json
