--- src/Bit.ComponentsPlayground.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32611.2
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bit.ComponentsPlayground.Server", "Server\Bit.ComponentsPlayground.Server.csproj", "{541882A1-1C06-48A9-8621-BC2C494F335F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bit.ComponentsPlayground.Client", "client\Bit.ComponentsPlayground.Client.csproj", "{7A15D250-904E-433B-9A0E-224D7E9615CD}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Bit.ComponentsPlayground.Shared", "Shared\Bit.ComponentsPlayground.Shared.csproj", "{81E14A0D-098E-4145-A707-BD5DA8199D0A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{541882A1-1C06-48A9-8621-BC2C494F335F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{541882A1-1C06-48A9-8621-BC2C494F335F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{541882A1-1C06-48A9-8621-BC2C494F335F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{541882A1-1C06-48A9-8621-BC2C494F335F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7A15D250-904E-433B-9A0E-224D7E9615CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7A15D250-904E-433B-9A0E-224D7E9615CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7A15D250-904E-433B-9A0E-224D7E9615CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7A15D250-904E-433B-9A0E-224D7E9615CD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{81E14A0D-098E-4145-A707-BD5DA8199D0A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{81E14A0D-098E-4145-A707-BD5DA8199D0A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{81E14A0D-098E-4145-A707-BD5DA8199D0A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{81E14A0D-098E-4145-A707-BD5DA8199D0A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5732240A-B642-4F05-955B-43638A6AF6B6}
-	EndGlobalSection
-EndGlobal
