--- src/client/Bit.ComponentsPlayground.Client.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.BlazorWebAssembly">
-
-	<PropertyGroup>
-		<TargetFramework>net7.0</TargetFramework>
-		<Nullable>enable</Nullable>
-		<ImplicitUsings>enable</ImplicitUsings>
-	</PropertyGroup>
-
-	<ItemGroup>
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="7.0.2" />
-		<PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="7.0.2" PrivateAssets="all" />
-	</ItemGroup>
-
-	<ItemGroup>
-		<ProjectReference Include="..\Shared\Bit.ComponentsPlayground.Shared.csproj" />
-		<PackageReference Include="Bit.BlazorUI" Version="4.9.2-pre-01" />
-		<Using Include="Bit.BlazorUI" />
-	</ItemGroup>
-
-</Project>
