--- src/client/Pages/Issues/2188/Index.razor
@@ -1,20 +1,0 @@
-﻿@page "/issues/2188"
-
-@using System.Globalization
-@using System.Reflection
-
-<PageTitle>Bit Components Playground</PageTitle>
-
-<BitDatePicker @bind-Value="SelectedDate" GoToToday="برو به امروز" Style="width: 300px" FormatDate="yyyy/MM/dd hh:mm tt" Culture="Bit.ComponentsPlayground.Client.Pages.CultureInfoHelper.GetFaIrCultureByFarsiNames()"></BitDatePicker>
-
-<BitDatePicker @bind-Value="SelectedDate" GoToToday="برو به امروز" Style="width: 300px" FormatDate="yyyy/MM/dd hh:mm tt" Culture="Bit.ComponentsPlayground.Client.Pages.CultureInfoHelper.GetFaIrCultureByFingilishNames()"></BitDatePicker>
-
-<BitDatePicker @bind-Value="SelectedDate" GoToToday="برو به امروز" Style="width: 300px" FormatDate="yyyy/MM/dd hh:mm tt" Culture="Bit.ComponentsPlayground.Client.Pages.AppCultureInfoHelper.GetFaIrCultureByFarsiNames()"></BitDatePicker>
-
-<hr />
-
-@SelectedDate?.Date.ToLongDateString()
-
-@code {
-    DateTimeOffset? SelectedDate = new DateTime(1992, 8, 7);
-}
