--- src/client/Pages/Issues/2817/Index.razor
@@ -1,38 +1,0 @@
-﻿@page "/issues/2817"
-
-<PageTitle>Bit Components Playground</PageTitle>
-
-<div>
-    <BitCarousel Style="height: 100px" VisibleItemsCount="3" ScrollItemsCount="1" InfiniteSliding="false">
-        <BitCarouselItem>
-            <div>1</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>2</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>3</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>4</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>5</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>6</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>7</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>8</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>9</div>
-        </BitCarouselItem>
-        <BitCarouselItem>
-            <div>10</div>
-        </BitCarouselItem>
-    </BitCarousel>
-</div>
