--- src/client/Shared/MainLayout.razor
@@ -1,3 +1,0 @@
-﻿@inherits LayoutComponentBase
-
-@Body
