--- src/client/Shared/MainLayout.razor.css
@@ -1,1 +1,0 @@
-﻿
