--- src/client/_Imports.razor
@@ -1,11 +1,0 @@
-﻿@using System.Net.Http
-@using System.Net.Http.Json
-@using Microsoft.AspNetCore.Components.Forms
-@using Microsoft.AspNetCore.Components.Routing
-@using Microsoft.AspNetCore.Components.Web
-@using Microsoft.AspNetCore.Components.Web.Virtualization
-@using Microsoft.AspNetCore.Components.WebAssembly.Http
-@using Microsoft.JSInterop
-@using Bit.ComponentsPlayground.Client
-@using Bit.ComponentsPlayground.Client.Shared
-@using Bit.BlazorUI
