--- src/client/wwwroot/app.css
@@ -1,4 +1,0 @@
-﻿.sample-button {
-    margin: auto;
-    display: block;
-}
