--- src/client/wwwroot/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    <title>Bit Components Playground</title>
-    <base href="/" />
-    <link href="_content/Bit.BlazorUI/styles/bit.blazorui.min.css" rel="stylesheet" />
-    <link href="app.css" rel="stylesheet" />
-    <link href="Bit.ComponentsPlayground.Client.styles.css" rel="stylesheet" />
-</head>
-
-<body>
-    <div id="app">Loading...</div>
-    <script src="_content/Bit.BlazorUI/scripts/bit.blazorui.min.js"></script>
-    <script src="_framework/blazor.webassembly.js"></script>
-</body>
-
-</html>
