+++ src/global.json
@@ -1,0 +1,6 @@
+{
+    "sdk": {
+        "version": "7.0.102",
+        "rollForward": "disable"
+    }
+}
