--- src/server/Bit.ComponentsPlayground.Server.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.Server" Version="7.0.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\client\Bit.ComponentsPlayground.Client.csproj" />
-    <ProjectReference Include="..\shared\Bit.ComponentsPlayground.Shared.csproj" />
-  </ItemGroup>
-
-
-</Project>
