--- src/server/Program.cs
@@ -1,30 +1,0 @@
-using Microsoft.AspNetCore.ResponseCompression;
-
-var builder = WebApplication.CreateBuilder(args);
-
-builder.WebHost.UseUrls("http://localhost:5000");
-
-builder.Services.AddControllersWithViews();
-builder.Services.AddRazorPages();
-
-var app = builder.Build();
-
-if (app.Environment.IsDevelopment())
-{
-    app.UseWebAssemblyDebugging();
-}
-else
-{
-    app.UseHsts();
-}
-
-app.UseBlazorFrameworkFiles();
-app.UseStaticFiles();
-
-app.UseRouting();
-
-app.MapRazorPages();
-app.MapControllers();
-app.MapFallbackToFile("index.html");
-
-app.Run();
